style: 역할이 더 잘 드러나도록 클래스 이름 수정
