+++ docs/README.md
@@ -2,21 +2,21 @@
 
 ## 기능 목록
 
-- [x] **각 자동차에 이름을 부여한다.** -`RacingCarName #inputRacingCarName`
+- [x] **각 자동차에 이름을 부여한다.** -`InputRacingCarName #inputRacingCarName`
   - [x] 자동차 이름은 쉼표(,)를 기준으로 구분한다.
-  - [x] **예외처리 : 이름은 5자 이하만 가능하다.** -`CarList #IllegalArgumentException`
-- [x] **사용자는 몇 번의 이동을 할 것인지 입력한다.** -`RacingCount #inputRacingCount`
+  - [x] **예외처리 : 이름은 5자 이하만 가능하다.** -`Exception #IllegalArgumentException`
+- [x] **사용자는 몇 번의 이동을 할 것인지 입력한다.** -`InputRacingCount #inputRacingCount`
 
 
-- [x] **사용자가 입력한 이동 횟수만큼 경주를 반복한다.** -`Racing #racingMoveCount`
-  - [x] **한 번 이동할 때마다 각 자동차의 이동횟수를 출력한다.** -`Racing #printRacingCar`
-  - [x] **경주를 진행하는 자동차 개수만큼 반복한다.** -`ForwardCar #racingCarCount`
-    - [x] **0에서 9까지의 숫자 중 무작위 값을 구한다.** -`ForwardCar #getRandomNumber`
-    - [x] **만약 무작위 값이 4 이상이라면 해당 자동차에 1을 더한다.** -`ForwardCar #goControlStatement`
+- [x] **사용자가 입력한 이동 횟수만큼 경주를 반복한다.** -`RacingStatus #racingMoveCount`
+  - [x] **한 번 이동할 때마다 각 자동차의 이동횟수를 출력한다.** -`PrintRacingRankStatus #printRacingCar`
+  - [x] **경주를 진행하는 자동차 개수만큼 반복한다.** -`ForwardCarCondition #racingCarCount`
+    - [x] **0에서 9까지의 숫자 중 무작위 값을 구한다.** -`ForwardCarCondition #getRandomNumber`
+    - [x] **만약 무작위 값이 4 이상이라면 해당 자동차에 1을 더한다.** -`ForwardCarCondition #goControlStatement`
 
 
 - [x] **누가 우승했는지 알려준다.** -`Referee #callWinnerCar`
-  - [x] 가장 많이 전진한 차가 우승자다.
+  - [x] **가장 많이 전진한 차가 우승자다.** -`Referee #winnerCarSelect`
   - [x] 우승자가 여러 명일 경우 쉼표(,)를 이용하여 구분한다.
