+++ src/main/java/racingcar/domain/InputRacingCarName.java
@@ -4,7 +4,7 @@ import camp.nextstep.edu.missionutils.Console;
 import java.util.Arrays;
 import java.util.List;
 
-public class RacingCarName {
+public class InputRacingCarName {
     public List<String> inputRacingCarName() {
         System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
         String CarNames = Console.readLine();
